requirements.txt
Django==3.1
djangorestframework==3.11.0
psycopg2-binary>=2.8
GDAL==2.4.0
djangorestframework_simplejwt==4.6.0